refactor: move types of form context in separate files
